+++ src/main/java/com/sipios/refactoring/models/dtos/CustomerType.java
@@ -1,0 +1,5 @@
+package com.sipios.refactoring.models.dtos;
+
+public enum CustomerType {
+    STANDARD_CUSTOMER,PREMIUM_CUSTOMER,PLATINUM_CUSTOMER
+}
